adecuacion de las vistas de appointment y segun estado de pago de MercadoLibre
