+++ front/src/components/Calendar/Calendar.jsx
@@ -140,7 +140,7 @@ const Calendar = ({id, calendar}) => {
                             bgcolor: "#a7d5ec ",
                             color: "#0752df"
                           }}>
-                            <Link style={{ textDecoration: 'none', color: "#0752df"}} to={`/turno/${id}/${dia.fecha}/${turno.hora}`}>{`${hh}:${mm}`}</Link>  
+                            <Link style={{ textDecoration: 'none', color: "#0752df"}} to={`/turno/${id}/${dia.fecha}/${turno.hora}/reserva/0`}>{`${hh}:${mm}`}</Link>  
                             {/* <Typography>{`${hh}:${mm}`}</Typography> */}
                         </ListItemButton> : 
                         <ListItemButton disabled
